fix(插件扫描): 逐 bundle 子进程隔离，修复启动闪退
- 根因：Maschine 3（NI/Qt）在后台线程加载时初始化 QApplication，
  Qt 要求主线程（dispatch_assert_queue 断言直接杀进程）
- 扫描改为两段：主进程纯文件系统收集 bundle → 每个 bundle 一个子进程加载
  （子进程自身主线程满足插件要求；崩溃/挂起只影响该插件）
- 4 个并发 worker + 流式进度（列表逐步出现，扫描中不卡界面）
- 加载失败的 bundle 不再静默消失：保留为带 error 的占位条目，
  选择器灰色显示 + hover 显示失败原因
- yinhe-clap/yinhe-vst3：暴露 collect_bundles（纯文件系统，不加载）
  与 scan_bundle（单包扫描）
- 实测：Maschine 3 子进程扫描成功；主程序启动 32 秒稳定不闪退

+++ crates/yinhe-vst3/src/stream.rs
@@ -119,11 +119,13 @@ impl IBStreamTrait for MemoryStream {
         let mut inner = self.inner.lock().unwrap_or_else(|e| e.into_inner());
         let len = inner.data.len() as i64;
         let mode = mode as u32;
-        let base = if mode == kIBSeekSet {
+        // 常量类型随平台绑定不同（i32/u32），统一转换比较。
+        #[allow(clippy::unnecessary_cast)]
+        let base = if mode == kIBSeekSet as u32 {
             0
-        } else if mode == kIBSeekCur {
+        } else if mode == kIBSeekCur as u32 {
             inner.pos as i64
-        } else if mode == kIBSeekEnd {
+        } else if mode == kIBSeekEnd as u32 {
             len
         } else {
             return kInvalidArgument;
